+++ train.lua
@@ -23,6 +23,7 @@ modelClass.declareOpts(cmd)
 onmt.train.Optim.declareOpts(cmd)
 onmt.train.Trainer.declareOpts(cmd)
 onmt.train.Checkpoint.declareOpts(cmd)
+onmt.data.SampledDataset.declareOpts(cmd)
 
 cmd:text('')
 cmd:text('**Other options**')
@@ -66,7 +67,12 @@ local function main()
     os.exit(0)
   end
 
-  local trainData = onmt.data.Dataset.new(dataset.train.src, dataset.train.tgt)
+  local trainData
+  if opt.sample > 0 then
+     trainData = onmt.data.SampledDataset.new(dataset.train.src, dataset.train.tgt, opt)
+  else
+     trainData = onmt.data.Dataset.new(dataset.train.src, dataset.train.tgt)
+  end
   local validData = onmt.data.Dataset.new(dataset.valid.src, dataset.valid.tgt)
 
   trainData:setBatchSize(opt.max_batch_size)
@@ -106,6 +112,10 @@ local function main()
       model = themodel
     end
   end)
+
+  if opt.sample > 0 then
+    trainData:checkModel(model)
+  end
 
   -- Define optimization method.
   local optimStates = (checkpoint.info and checkpoint.info.optimStates) or nil